It works
Working directory was fout waardoor .xml niet juist werd ingeladen
