--- Testing/agentconfig.xml
@@ -1,3 +1,0 @@
-<config>
-	<port>8008</port>
-</config>
