+++ management-Jasper-PC1.csv
@@ -10,3 +10,6 @@
 2016-01-10 20:58:24,Jasper-PC1,Requested zooi,<resultaat><loggedinusers/><freespace/><uptime/><ip/><ram/><platform>win32</platform><services/></resultaat>
 2016-01-10 20:59:04,Jasper-PC1,Requested zooi,<resultaat><loggedinusers/><freespace/><uptime/><ip/><ram/><platform>win32</platform><services/></resultaat>
 2016-01-10 20:59:17,Jasper-PC1,Requested zooi,<resultaat><loggedinusers/><freespace/><uptime/><ip/><ram/><platform>win32</platform><services/></resultaat>
+2016-01-11 10:58:41,Jasper-PC1,Requested zooi,"<resultaat><loggedinusers>1</loggedinusers><freespace>111.99 GB</freespace><uptime>10 days, 13 hours, 34 minutes</uptime><ip>169.254.64.212</ip><ram>16266 11030</ram><platform>win32</platform><services>112 running services en 121 gestopte</services></resultaat>"
+2016-01-11 11:01:33,Jasper-PC1,Requested zooi,"<resultaat><loggedinusers>1</loggedinusers><freespace>111.99 GB</freespace><uptime>10 days, 13 hours, 37 minutes</uptime><ip>169.254.64.212</ip><ram>16266 10960</ram><platform>win32</platform><services>111 running services en 122 gestopte</services></resultaat>"
+2016-01-11 11:03:37,Jasper-PC1,Requested zooi,"<resultaat><loggedinusers>1</loggedinusers><freespace>111.99 GB</freespace><uptime>10 days, 13 hours, 39 minutes</uptime><ip>169.254.64.212</ip><ram>16266 10930</ram><platform>win32</platform><services>112 running services en 121 gestopte</services></resultaat>"
