--- management-sjabloon.csv
@@ -1,1 +1,0 @@
-Datumtijd,Agent,Request,Response
