+++ mgmtconfig.xml
@@ -1,6 +1,15 @@
+<?xml version="1.0" encoding="ISO-8859-1"?>
 <config>
-	<hostname>127.0.0.1</hostname>
-	<port>8008</port>
-      <logfile>management.log</logfile>
-      <loglevel>WARNING</loglevel>
+	<server name="Server1"> <!-- name is weergave naam -->
+		<host>localhost</host><!-- Mag IP of hostname zijn-->
+		<port>8008</port> <!-- Port waarop de server draait -->
+	</server>
+	<server name="Server2"> <!-- name is weergave naam -->
+		<host>localhost</host><!-- Mag IP of hostname zijn-->
+		<port>8009</port> <!-- Port waarop de server draait -->
+	</server>
+	<logging>
+		<logfile>management.log</logfile>
+		<loglevel>WARNING</loglevel>
+	</logging>
 </config>
